Add Hashnode's Expo community link
Expo community on Hashnode has started growing in size recently. It would be great if this community is linked from the official docs.

<!--
Thanks for helping! Please check that you have edited the docs in the
`versions/unversioned` directory, if you want these changes to apply to
the next SDK version too.
-->

Closes expo/expo-docs#304

fbshipit-source-id: 3e7e19f